+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Add string case support for `$toDate` operator
 * Support collation with `$indexOfArray` aggregation operator. *[Feature]*
 * Support collation with arrays and objects comparisons *[Feature]*
+* Support background index builds *[Bugfix]* (#36)
 
 ### documentdb v0.103-0 (Unreleased) ###
 * Support collation with aggregation and find on sharded collections *[Feature]*